Moved food security calculation to dashboard
